Properly Name Auto, adds functionality to pass in wait time
renamed "OneBallandSteal" to "OneBallandD" because no stealing actually happens. there was previously a path named "OneBallandD" however should it run it would break so it was removed. 1BallAndD, LeftTrickTaxi, and RightTrickTaxi all have the capability to pass in a value for wait time, it is not required but can be passed in.